requirements.txt
google-cloud-bigquery>=3.40.0
apache-airflow>=2.8.0
